scripts/hybrid/requirements.txt
# Superset Hybrid Control Plane - Python Dependencies

# Core dependencies
Jinja2>=3.1.0
PyYAML>=6.0
requests>=2.31.0

# Validation and quality
yamllint>=1.33.0
jsonschema>=4.20.0

# Superset API integration
apache-superset>=3.0.0  # For API client libraries

# Database connectivity (for drift detection)
psycopg2-binary>=2.9.0  # PostgreSQL (Supabase)
SQLAlchemy>=2.0.0

# Testing and quality assurance
pytest>=7.4.0
pytest-cov>=4.1.0

# Visual parity testing
playwright>=1.40.0
scikit-image>=0.22.0
numpy>=1.24.0
Pillow>=10.0.0

# CLI enhancements
click>=8.1.0
rich>=13.0.0  # For beautiful terminal output
tabulate>=0.9.0

# Environment management
python-dotenv>=1.0.0
